Exception handling - delete

+++ src/main/java/com/felipelourenco/course/services/UserService.java
@@ -2,8 +2,12 @@ package com.felipelourenco.course.services;
 
 import com.felipelourenco.course.entities.User;
 import com.felipelourenco.course.repositories.UserRepository;
+import com.felipelourenco.course.services.exceptions.DatabaseException;
 import com.felipelourenco.course.services.exceptions.ResourceNotFoundException;
+import org.apache.catalina.webresources.EmptyResource;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +33,14 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        }catch (ResourceNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user) {
